Set up zppy executable
