+++ zppy/templates/default.ini
@@ -83,7 +83,6 @@ output_format_subplot = string_list(default=list(""))
 multiprocessing = boolean(default=True)
 num_workers = integer(default=24)
 vars = string(default="FSNTOA,FLUT,FSNT,FLNT,FSNS,FLNS,SHFLX,QFLX,TAUX,TAUY,PRECC,PRECL,PRECSC,PRECSL,TS,TREFHT,CLDTOT,CLDHGH,CLDMED,CLDLOW,U")
-environment_commands = string(default="")
 
   [[__many__]]
   active = boolean(default=None)
@@ -104,7 +103,6 @@ environment_commands = string(default="")
   multiprocessing = boolean(default=None)
   num_workers = integer(default=None)
   vars = string(default=None)
-  environment_commands = string(default=None)
   climo_diurnal_subsection = string(default=None)
 
 [e3sm_diags_vs_model]
