+++ zppy/templates/e3sm_diags/lat_lon_conus.cfg
